feat(status): setup response from post
