+++ yaki_mobile/lib/data/repositories/login_repository.dart
@@ -1,3 +1,4 @@
+import 'package:yaki/data/models/authentication.dart';
 import 'package:yaki/data/sources/login_api.dart';
 import 'package:yaki/data/models/login.dart';
 
@@ -6,8 +7,8 @@ class LoginRepository {
 
   LoginRepository(this._loginService);
 
-  Future<Login?> postLogin(Login login) {
-    final log = _loginService.postLogin(login);
+  Future<Authentication?> postLogin(Login login) async {
+    final log = await _loginService.postLogin(login);
     return log;
   }
 }
