+++ yaki_mobile/lib/data/sources/login_api.dart
@@ -1,5 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
+import 'package:yaki/data/models/authentication.dart';
 import 'package:yaki/data/models/login.dart';
 
 part 'login_api.g.dart';
@@ -9,5 +10,5 @@ abstract class LoginService {
   factory LoginService(Dio dio, {required String baseUrl}) = _LoginService;
 
   @POST('/login')
-  Future<Login?> postLogin(@Body() Login login);
+  Future<Authentication?> postLogin(@Body() Login login);
 }
