+++ yaki_mobile/lib/presentation/state/notifiers/declaration_notifier.dart
@@ -36,7 +36,8 @@ class DeclarationNotifier extends StateNotifier<dynamic> {
     setState(status, true);
 
     final statusCreateResponse = await declarationRepository.create(newDeclaration);
+    print("test");
+    print('print declaratio response ${statusCreateResponse.toJson()}');
 
-    // need testing to check received answer.
   }
 }
