+++ yaki_mobile/lib/presentation/state/notifiers/login_notifier.dart
@@ -17,6 +17,11 @@ class LoginNotifier extends StateNotifier<String> {
       salt: const String.fromEnvironment('CRED_HASH_PASS'),
     );
     final newLog = Login(login: newLogin, password: hashPass.toString());
-    repository.postLogin(newLog);
+    print('log object: ${newLog.toJson()}');
+    final createResp = await repository.postLogin(newLog);
+
+    print('login response ${createResp?.toJson()}');
+
+
   }
 }
