+++ yaki_mobile/lib/presentation/ui/declaration/views/declaration_body.dart
@@ -4,7 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:yaki/presentation/displaydata/declaration_card_content.dart';
 import 'package:yaki/presentation/state/providers/provider.dart';
-import 'package:yaki/presentation/ui/declaration/views/status_card.dart';
+import 'package:yaki/presentation/ui/declaration/views/status_card_test_media.dart';
 
 
 void _routeHandling(BuildContext context) {
