+++ yaki_mobile/lib/presentation/ui/status/status.dart
@@ -14,6 +14,7 @@ void _routeHandling(BuildContext context) {
 }
 void onAvatarIconPress() {}
 
+
 class Status extends ConsumerWidget {
   const Status({Key? key}) : super(key: key);
